Begin dark theme skeleton for mobile style
So far this coveres the article, talk, history, and recent changes
pages, as well as most common UI elements like search, sidebar, and
most of the menus.

Some of the rules are lifted straight from the non-mobile style. In
the future, these common rules could be extracted out into a common
section to reduce duplication.