Remove generic `div` selector, and fix fallout
We should not be using generic selectors with !important, as this
inevitably results in having !important spammed all over the stylesheet.

This patch is a first attempt to remove the most problematic one (div),
which required several new selectors to take its place.

The style post-change should have parity with the previous version.