Removal of problematic code
A segment of code is prohibiting custom colours from showing, which is problematic for when one wishes to see colour previews from the wiki's pages about colours. After reviewing the code segment in question, it appears like a cobbled-together mess that does not offer a lot of room for dynamic situations. If you wish to add back the sub-table grey, then I implore you to reconsider the coding logic before doing so.